routing key configuration

+++ suro-client/src/main/java/com/netflix/suro/message/MessageSetBuilder.java
@@ -70,7 +70,7 @@ public class MessageSetBuilder {
             return new TMessageSet(
                     config.getApp(),
                     messageList.size(),
-                    compression.getId(),
+                    (byte) compression.getId(),
                     crc,
                     ByteBuffer.wrap(buffer));
         } catch (IOException e) {
